+++ scanpipe/tests/test_commands.py
@@ -119,16 +119,16 @@ class ScanPipeManagementCommandTest(TestCase):
     def test_scanpipe_management_command_create_project_inputs(self):
         out = StringIO()
 
-        options = ["--input", "non-existing"]
+        options = ["--input-file", "non-existing"]
         expected = "non-existing not found or not a file"
         with self.assertRaisesMessage(CommandError, expected):
             call_command("create-project", "my_project", *options)
 
         parent_path = Path(__file__).parent
         options = [
-            "--input",
+            "--input-file",
             str(parent_path / "test_commands.py"),
-            "--input",
+            "--input-file",
             str(parent_path / "test_models.py"),
         ]
         call_command("create-project", "my_project", *options, stdout=out)
@@ -158,13 +158,15 @@ class ScanPipeManagementCommandTest(TestCase):
         self.assertIn(f"Pipeline {pipeline} run in progress...", out.getvalue())
         self.assertIn("successfully executed on project my_project", out.getvalue())
 
-    def test_scanpipe_management_command_add_input(self):
+    def test_scanpipe_management_command_add_input_file(self):
         out = StringIO()
 
         project = Project.objects.create(name="my_project")
         parent_path = Path(__file__).parent
         options = [
+            "--input-file",
             str(parent_path / "test_commands.py"),
+            "--input-file",
             str(parent_path / "test_models.py"),
         ]
 
@@ -178,10 +180,22 @@ class ScanPipeManagementCommandTest(TestCase):
         expected = sorted(["test_commands.py", "test_models.py"])
         self.assertEqual(expected, sorted(project.input_files))
 
-        options = ["--project", project.name, "non-existing.py"]
+        options = ["--project", project.name, "--input-file", "non-existing.py"]
         expected = "non-existing.py not found or not a file"
         with self.assertRaisesMessage(CommandError, expected):
             call_command("add-input", *options, stdout=out)
+
+    def test_scanpipe_management_command_add_input_url(self):
+        out = StringIO()
+
+        project = Project.objects.create(name="my_project")
+        parent_path = Path(__file__).parent
+        options = [
+            "--input-file",
+            str(parent_path / "test_commands.py"),
+            "--input-file",
+            str(parent_path / "test_models.py"),
+        ]
 
     def test_scanpipe_management_command_add_pipeline(self):
         out = StringIO()