+++ scanpipe/tests/test_models.py
@@ -25,6 +25,7 @@ import tempfile
 import uuid
 from contextlib import redirect_stdout
 from datetime import datetime
+from operator import attrgetter
 from pathlib import Path
 from unittest import mock
 
@@ -39,6 +40,8 @@ from scanpipe.models import DiscoveredPackage
 from scanpipe.models import Project
 from scanpipe.models import ProjectError
 from scanpipe.models import Run
+from scanpipe.pipes.fetch import Download
+from scanpipe.pipes.input import copy_inputs
 from scanpipe.tests import mocked_now
 from scanpipe.tests import package_data1
 from scanpipe.tests.pipelines.do_nothing import DoNothing
@@ -90,7 +93,8 @@ class ScanPipeModelsTest(TestCase):
         work_path = self.project1.work_path
         self.assertTrue(work_path.exists())
 
-        self.project1.add_input_file(SimpleUploadedFile("file.ext", content=b"content"))
+        uploaded_file = SimpleUploadedFile("file.ext", content=b"content")
+        self.project1.write_input_file(uploaded_file)
         self.project1.add_pipeline("docker")
         resource = CodebaseResource.objects.create(project=self.project1, path="path")
         package = DiscoveredPackage.objects.create(project=self.project1)
@@ -137,11 +141,11 @@ class ScanPipeModelsTest(TestCase):
         filename = self.project1.get_output_file_path("file", "ext")
         self.assertTrue(str(filename).endswith("/output/file-2010-10-10-10-10-10.ext"))
 
-    def test_scanpipe_project_model_add_input_file(self):
+    def test_scanpipe_project_model_write_input_file(self):
         self.assertEqual([], self.project1.input_files)
 
         uploaded_file = SimpleUploadedFile("file.ext", content=b"content")
-        self.project1.add_input_file(uploaded_file)
+        self.project1.write_input_file(uploaded_file)
 
         self.assertEqual(["file.ext"], self.project1.input_files)
 
@@ -164,6 +168,98 @@ class ScanPipeModelsTest(TestCase):
         self.project1.move_input_from(input_location)
         self.assertEqual([input_filename], self.project1.input_files)
         self.assertFalse(Path(input_location).exists())
+
+    def test_scanpipe_project_model_inputs_with_source(self):
+        inputs, missing_inputs = self.project1.inputs_with_source
+        self.assertEqual([], inputs)
+        self.assertEqual({}, missing_inputs)
+
+        uploaded_file = SimpleUploadedFile("file.ext", content=b"content")
+        self.project1.add_uploads([uploaded_file])
+        self.project1.copy_input_from(self.data_location / "notice.NOTICE")
+        self.project1.add_input_source(filename="missing.zip", source="uploaded")
+
+        inputs, missing_inputs = self.project1.inputs_with_source
+        expected = [
+            {
+                "is_file": True,
+                "name": "file.ext",
+                "size": 7,
+                "source": "uploaded",
+            },
+            {
+                "is_file": True,
+                "name": "notice.NOTICE",
+                "size": 1178,
+                "source": "not_found",
+            },
+        ]
+
+        def sort_by_name(x):
+            return x.get("name")
+
+        self.assertEqual(
+            sorted(expected, key=sort_by_name), sorted(inputs, key=sort_by_name)
+        )
+        self.assertEqual({"missing.zip": "uploaded"}, missing_inputs)
+
+    def test_scanpipe_project_model_can_add_input(self):
+        self.assertTrue(self.project1.can_add_input)
+
+        run = self.project1.add_pipeline("docker")
+        self.project1 = Project.objects.get(uuid=self.project1.uuid)
+        self.assertTrue(self.project1.can_add_input)
+
+        run.task_start_date = timezone.now()
+        run.save()
+        self.project1 = Project.objects.get(uuid=self.project1.uuid)
+        self.assertFalse(self.project1.can_add_input)
+
+    def test_scanpipe_project_model_add_input_source(self):
+        self.assertEqual({}, self.project1.input_sources)
+
+        self.project1.add_input_source("filename", "source", save=True)
+        self.project1.refresh_from_db()
+        self.assertEqual({"filename": "source"}, self.project1.input_sources)
+
+    def test_scanpipe_project_model_add_downloads(self):
+        file_location = self.data_location / "notice.NOTICE"
+        copy_inputs([file_location], self.project1.tmp_path)
+
+        download = Download(
+            uri="https://example.com/filename.zip",
+            directory="",
+            filename="notice.NOTICE",
+            path=self.project1.tmp_path / "notice.NOTICE",
+            size="",
+            sha1="",
+            md5="",
+        )
+
+        self.project1.add_downloads([download])
+
+        inputs, missing_inputs = self.project1.inputs_with_source
+        expected = [
+            {
+                "is_file": True,
+                "name": "notice.NOTICE",
+                "size": 1178,
+                "source": "https://example.com/filename.zip",
+            }
+        ]
+        self.assertEqual(expected, inputs)
+        self.assertEqual({}, missing_inputs)
+
+    def test_scanpipe_project_model_add_uploads(self):
+        uploaded_file = SimpleUploadedFile("file.ext", content=b"content")
+        self.project1.add_uploads([uploaded_file])
+
+        inputs, missing_inputs = self.project1.inputs_with_source
+        expected = [
+            {"name": "file.ext", "is_file": True, "size": 7, "source": "uploaded"}
+        ]
+        self.assertEqual(expected, inputs)
+        self.assertEqual({}, missing_inputs)
 
     def test_scanpipe_project_model_get_next_run(self):
         self.assertEqual(None, self.project1.get_next_run())